requirements.txt
# Editable install with no version control (agent==0.0.1)
-e /home/abdulrafay/Desktop/str_agent/support-agent
annotated-types==0.7.0
anyio==4.9.0
blockbuster==1.5.25
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cryptography==44.0.3
filelock==3.18.0
forbiddenfruit==0.1.4
fsspec==2025.7.0
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.26
langchain-core==0.3.69
langchain-text-splitters==0.3.8
langgraph==0.5.3
langgraph-api==0.2.98
langgraph-checkpoint==2.1.1
langgraph-cli==0.3.5
langgraph-prebuilt==0.5.2
langgraph-runtime-inmem==0.6.0
langgraph-sdk==0.1.73
langsmith==0.4.8
orjson==3.11.0
ormsgpack==1.10.0
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.1.3
starlette==0.47.1
structlog==25.4.0
# Editable install with no version control (support-agent==0.0.1)
-e /home/abdulrafay/Desktop/str_agent/support-agent
tenacity==9.1.2
tqdm==4.67.1
truststore==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
xxhash==3.5.0
zstandard==0.23.0
